# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SOLOS+ ESPORTZ clan management platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "401":
          description: Unauthorized

  # ── MEMBERS ───────────────────────────────────────────────────────────────
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List all announcements
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"

    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement (management only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    get:
      operationId: getAnnouncement
      tags: [announcements]
      summary: Get announcement by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"

    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete announcement (management only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── SCRIMS ────────────────────────────────────────────────────────────────
...
  /chat/groups/{groupId}/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get messages for a chat group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"

    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message to a chat group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
  /chat/groups/{groupId}/messages/{messageId}:
    delete:
      operationId: deleteMessage
      tags: [chat]
      summary: Delete a message (management or own)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── NOTIFICATIONS ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Member:
      type: object
      required: [id, email, codmUsername, displayName, role, status, clanPoints, kdRatio, mvpCount, activityScore, scrimWins, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        codmUsername:
          type: string
        displayName:
          type: string
          description: "Clan-tagged username e.g. S²十NESS"
        whatsappNumber:
          type: ["string", "null"]
        customTag:
          type: ["string", "null"]
          description: "Custom role tag e.g. Admin, Clan Master, Sniper"
        role:
          type: string
          enum: [OWNER, MANAGEMENT, TIER1, TIER2, TIER3, NEW_MEMBER]
        status:
          type: string
          enum: [pending, active, rejected, kicked]
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        clanPoints:
          type: integer
        kdRatio:
          type: number
        mvpCount:
          type: integer
        activityScore:
          type: integer
        scrimWins:
          type: integer
        isOnline:
          type: boolean
        lastSeen:
          type: ["string", "null"]
        achievements:
          type: array
          items:
            type: string
        createdAt:
          type: string
...
    Announcement:
      type: object
      required: [id, title, content, type, authorId, authorName, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [general, scrim, tournament, promotion, meeting, urgent]
        authorId:
          type: string
        authorName:
          type: string
        pinned:
          type: boolean
        createdAt:
          type: string
    AnnouncementInput:
      type: object
      required: [title, content, type]
      properties:
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [general, scrim, tournament, promotion, meeting, urgent]
        pinned:
          type: boolean
...
    ChatMessage:
      type: object
      required: [id, groupId, authorId, authorName, authorRole, content, createdAt]
      properties:
        id:
          type: string
        groupId:
          type: string
        authorId:
          type: string
        authorName:
          type: string
        authorRole:
          type: string
        avatarUrl:
          type: ["string", "null"]
        content:
          type: string
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
...